[WIP] Refactored VicBackends in personality server
Moved all swagger specific code outside the main backend files.
This should make the main backend files a lot cleaner.  Also
refactored the swagger calls out of the API calls into new
functions in the _portlayer.go files.

Added rules comments for how to add code to both the main backend
files and the _portlayer.go files.
